MIPS: R5900: Initial support for the Emotion Engine in the PlayStation 2
The R5900 implements the 64-bit MIPS III instruction set except LL, SC,
LLD and SCD, with additional PREFETCH and conditional move instructions
from MIPS IV as well as three-operand multiply and multiply-accumulate
instructions. A set of 128-bit multimedia instructions specific to the
R5900 is also implemented. Endianness is not configurable and taken to
be little-endian. The R5900 does not implement CP0.Status.{UX,SX,KX,PX}.

The COP1 FPU implements single-precision floating-point operations but
is not entirely IEEE 754 compatible. In particular,

- NaN (not a number) and plus/minus infinities are not supported;
- exception mechanisms are not fully supported;
- denormalized numbers are not supported;
- rounding towards nearest and plus/minus infinities are not supported;
- computed results usually differs in the least significant bit;
- saturating instructions can differ more than the least significant bit.

Since only rounding towards zero is supported, the two least significant
bits of FCR31 are hardwired to 01. To support IEEE 754 in applications
the FPU is emulated in software by the kernel.

The VPU0 is a vector processor of the Emotion Engine. In macro mode, it
functions as COP2 (coprocessor) and instructions execute simultaneously
in the main integer pipeline I1 and the COP2 pipeline. In micro mode,
the VPU0 functions as a stand-alone processor. The VPU1 is an additional
vector processor that operates independently of both the R5900 and the
VPU0. It primarily acts as a preprocessor to the Graphics Synthesizer.

The scratch pad RAM (SPRAM) of the Emotion Engine is 16 KiB of very fast
static RAM organised in 128-bit quadwords. Both the DMA controller and
the R5900 can access the SPRAM.

The R5900 has several significant hardware bugs. Perhaps the most
important bug affecting applications is the short loop bug, which under
certain conditions causes loops to execute only once or twice. The Gnu
assembler has the following note:

    We must ensure that the assembler never generates loops that satisfy
    all of the following conditions:

    - a loop consists of less than equal to six instructions
      (including the branch delay slot);
    - a loop contains only one conditional branch instruction at
      the end of the loop;
    - a loop does not contain any other branch or jump instructions;
    - a branch delay slot of the loop is not NOP (EE 2.9 or later).

Signed-off-by: Fredrik Noring <noring@nocrew.org>