MIPS: R5900: Trap the RDHWR instruction as an SQ address exception
On the R5900, the RDHWR instruction is interpreted as the R5900 specific
SQ instruction[1] that traps into a zero page address exception. Hence
RDHWR can be emulated by emulate_load_store_insn().

CONFIG_DEFAULT_MMAP_MIN_ADDR must not be less than PAGE_SIZE to reliably
trap and emulate RDHWR, so this is made a BUILD_BUG_ON for the R5900.

References:

[1] "TX System RISC TX79 Core Architecture" manual, revision 2.0,
    Toshiba Corporation, p. B-162, https://wiki.qemu.org/File:C790.pdf

Signed-off-by: Fredrik Noring <noring@nocrew.org>