MIPS: R5900: Add the SYNC.P instruction
The SYNC.P instruction is a pipeline barrier. All instructions prior to
the barrier are completed before the instructions following the barrier
are fetched.

However, the barrier operation doesn't wait for any prior instructions
to retire, for example multiply, divide, multicycle COP1 operations or
a pending load issued before the barrier operation.

Signed-off-by: Fredrik Noring <noring@nocrew.org>