MIPS: R5900: Add the SYNC.P instruction
The SYNC.P instruction is a pipeline barrier. All instructions prior to
the barrier are completed before the instructions following the barrier
are fetched[1].

However, the barrier operation doesn't wait for any prior instructions
to retire, for example multiply, divide, multicycle COP1 operations or
a pending load issued before the barrier operation.

References:

[1] "TX System RISC TX79 Core Architecture" manual, revision 2.0,
    Toshiba Corporation, p. A-125, https://wiki.qemu.org/File:C790.pdf

Signed-off-by: Fredrik Noring <noring@nocrew.org>