MIPS: R5900: Add mandatory SYNC.P to all M[FT]C0 instructions
The Toshiba TX79 manual mandates that all MTC0 instructions must be
followed by a SYNC.P instruction as a barrier to guarantee COP0 register
updates. There is one exception to this rule:

An MTC0 instruction which loads the EntryHi COP0 register can be followed
by a TLBWI or a TLBWR instruction without having an intervening SYNC.P.
This special case is handled by a hardware interlock.

Signed-off-by: Fredrik Noring <noring@nocrew.org>