MIPS: R5900: Add MFSA and MTSA instructions for the special SA register
The shift amount (SA) register is a 64-bit special register storing the
funnel shift amount. The SA is encoded an implementation-defined manner.
It is therefore not meaningful for software to operate on this value.
Use the MTSAB and MTSAH instructions to set a new funnel shift amount.

The SA is used by the QFSRV (quadword funnel shift right variable)
256-bit multimedia instruction.

MFSA copies the SA register to a 64-bit GPR. The sole purpose of this
instruction is to permit the shift amount to be saved during a context
switch.

MTSA copies a 64-bit GPR rs to the SA register. Note that rs must
contain a value that was originally generated by MFSA. If some other
user-generated value is in rs, the shifting action performed by the
funnel shifter is not defined; that is, MTSA cannot be used to by a
program to set a new funnel shift amount. The sole purpose of this
instruction is to permit the shift amount to be restored during a
context switch. Restrictions:

The three instructions statically preceding a MTSA instruction must
not read or write the SA register; that is, they cannot be either of
the instructions MFSA, QFSRV, or MTSAx.

Signed-off-by: Fredrik Noring <noring@nocrew.org>