MIPS: PS2: FB: Frame buffer driver for the PlayStation 2
The main limitation is the lack of mmap, since the Graphics Synthesizer
has local frame buffer memory that is not directly accessible from the
main bus. The GS has 4 MiB of local memory.

The console drawing primitives are synchronous to allow printk at any
time. This is highly useful for debugging but it is not the fastest
possible implementation. The console is nevertheless very fast and
makes use of several hardware accelerated features of the Graphics
Synthesizer.

The maximum practical resolution is 1920x1080p at 16 bits per pixel that
requires 4147200 bytes of local memory, leaving 47104 bytes for a tiled
font, which at 8x8 pixels and a minimum 4 bits indexed texture palette is
at most 1464 characters. The indexed palette makes switching colours easy.
&struct fb_tile_ops is accelerated with GS texture sprites that are fast
(GS local copy) for the kernel via simple DMA GS commands via the GIF.

Signed-off-by: Fredrik Noring <noring@nocrew.org>