USB: OHCI: Support for the PlayStation 2
Support the USB 1.1 OHCI for the PlayStation 2. The controller is
connected to the I/O processor (IOP) which is separate from the main
Emotion Engine (EE) processor that runs the kernel.

IOP DMA memory and OHCI registers are mapped into EE address space.
OHCI interrupts are asserted on the IOP and relayed to the EE by an
IOP driver.

Signed-off-by: Fredrik Noring <noring@nocrew.org>