MIPS: PS2: Timer support
The Emotion Engine has 4 independent timers with 16-bit counters[1]. The
bus clock or an external (H-BLANK or V-BLANK) clock perform the counting.
When a counter reaches a specified reference value, or when it overflows,
an interrupt is asserted. The timer status register indicates the cause.

Timers 0 and 1 have hold registers for recording the counter value when
an SBUS interrupt occurs.

Timer registers are 32-bit long and only word-accessible.

References:

[1] "EE User's Manual", version 6.0, Sony Computer Entertainment Inc.,
    pp. 33-39.

Signed-off-by: Fredrik Noring <noring@nocrew.org>