USB: OHCI: Support for the PlayStation 2
Support the USB 1.1 OHCI for the PlayStation 2. The controller is
connected to the input/output processor (IOP) which is separate from
the main (R5900) processor that runs the kernel.

IOP DMA memory and OHCI registers are mapped into kernel address space.
OHCI interrupts are asserted on the IOP and forwarded to the kernel by
an IOP IRQ relay module.

Signed-off-by: Fredrik Noring <noring@nocrew.org>