MIPS: PS2: SIF: Respond to remote procedure call (RPC) bind command
The IOP can service remote procedure calls (RPCs) from the main (R5900)
processor. The first step is to send a bind command that the IOP will
respond to, to establish a client connection. The main processor responds
by issuing a bind end command to the IOP.

Allocating IOP memory and linking IOP modules are examples of IOP RPC
services that will be used in subsequent changes. IOP memory is needed
for USB OHCI support.

Signed-off-by: Fredrik Noring <noring@nocrew.org>