MIPS: PS2: SIF: sif_rpc() to issue a remote procedure call
This is the actual function to perform IOP procedure calls. For
simplicity a call is synchronous. It is certainly possible to make calls
more efficient, asynchronous, and so on, but ease of use is often more
important than performance, for example when allocating IOP memory
during initialisation of a driver.

Signed-off-by: Fredrik Noring <noring@nocrew.org>