MIPS: PS2: GIF: Define Graphics Synthesizer interface structures
The Graphics Synthesizer interface (GIF) formats data based on the GIF
tags at the start of display list packets, and then transfer formatted
data to the GS as drawing commands[1].

The GIF has three kinds of general data transfer paths:

- PATH1: data from vector processor unit 1 (VPU1) memory;
- PATH2: data via the VPU1 interface packet expansion engine (VIF1) FIFO;
- PATH3: data directly from the main bus to the Graphics Synthesizer.

The GIF tag is a 16-byte fixed length object that specifies the size
and structure of subsequent data. Up to 16 register descriptors can be
specified in one GIF tag, specifying to which registers of the GS the
input data following the GIF tag should be output and the packing
format mode:

- in PACKED mode, the data of each quadword is interpreted and packed
  according to the register descriptor in the GIF tag, and is output
  to the address specified in the same way by the register descriptor;

- in REGLIST mode, the data following the GIF tag is considered to be
  data strings of 2x64 bits and output as is without packing by setting
  the register descriptor value as the output destination address;

- in IMAGE mode, the data following the GIF tag is considered to be
  data strings of 2x64 bits and output to the host-local transfer
  register HWREG of the GS, used when transferring image data such as
  textures.

The privileged registers of the GS are directly mapped in the address
space of the main bus, and are not accessible via the GIF regardless of
the state of the general data transfer path. The GIF monitors access to
the privileged registers.

It is nontrivial to transfer data from GS local memory by reversing the
direction with the privileged BUSDIR register.

References:

[1] "EE User's Manual", version 6.0, Sony Computer Entertainment Inc.,
    pp. 147-160.

Signed-off-by: Fredrik Noring <noring@nocrew.org>