MIPS: PS2: FB: Preconfigure standard PAL, NTSC and VESA display modes
These display mode configurations are a combination of modes made by
Sony[1] and modes from the PlayStation 3 frame buffer driver (ps3fb.c).

The VESA modes are designed to work with the synch-on-green (SOG) VGA
cable that Sony distributed with the Linux kit for the PlayStation 2.
Modern computer displays typically do not support synch-on-green, so
an adapter is most likely necessary for these modes.

References:

[1] "GS User's Manual", version 6.0, Sony Computer Entertainment Inc.,
    p. 84.

Signed-off-by: Fredrik Noring <noring@nocrew.org>