MIPS: PS2: ROM: Read and decode the ROMVER file
The ROMVER file contains information on the ROM version, the machine
region, the machine type and the date the ROM was created.

The ROM region is 'J' for Japan, 'E' for Europe, 'C' for China, and 'A'
or 'H' for the USA; 'X' indicates a TEST machine, and 'T' either a TOOL
machine or a Namco system if the ROM type is 'Z'.

The ROM type is 'C' for retail (CEX), 'D' for debug (DEX), and 'Z' for
Namco systems.

All known ROMVER files:

0100JC20000117  0150EC20001228  0170EC20030227  0220AC20050620
0100TD20000117  0150ED20001228  0170ED20030227  0220AC20060210
0100TD20000117  0150JC20010118  0170JC20030206  0220AC20060905
0100TD20000125  0150JD20010118  0180CD20030224  0220AD20060905
0100TD20031031  0160AC20010427  0180JC20031028  0220EC20050620
0100TD20051124  0160AC20010704  0190AC20030623  0220EC20060210
0100TZ20000306  0160AC20011004  0190AC20040329  0220EC20060905
0101JC20000217  0160AC20020207  0190CC20030623  0220HC20050620
0101JD20000217  0160EC20010704  0190EC20030623  0220HC20060210
0101XD20000224  0160EC20011004  0190EC20030822  0220HC20060905
0110AC20000727  0160EC20020319  0190HC20030623  0220JC20050620
0110AD20000727  0160EC20020426  0190JC20030623  0220JC20060210
0120AC20000902  0160HC20010730  0190JC20030822  0220JC20060905
0120EC20000902  0160HC20020426  0200AC20040614  0220JD20121227
0120ED20000902  0160JC20010427  0200EC20040614  0230AC20080220
0120JC20001027  0160JC20020426  0200HC20040614  0230EC20080220
0150AC20001228  0170AC20030325  0200JC20040614  0230HC20080220
0150AD20001228  0170AD20030325  0210JC20040917  0250EC20100415

Signed-off-by: Fredrik Noring <noring@nocrew.org>