MIPS: PS2: ROM: Read and decode the ROMVER file
The ROMVER file contains information on the ROM version, the machine
region, the machine type (CEX for retail, DEX for debug, or TOOL) and
the date the ROM was created.

The machine region can be used to default to the PAL video mode for
European machines and NTSC for Japanese machines, for example.

Signed-off-by: Fredrik Noring <noring@nocrew.org>