MIPS: PS2: ROM: Sysfs module to inspect ROM files
The ROM sysfs module gives information on the ROM file names, sizes,
and addresses. For example, listing all files of ROM0:

	# cat /sys/rom/rom0/file/*/name | column
	RESET       VBLANK      RMRESET     MCSERV      XSIFCMD
	ROMDIR      IOMAN       OSDVER      PADMAN      XCDVDMAN
	EXTINFO     MODLOAD     -           CDVDMAN     XCDVDFSV
	ROMVER      ROMDRV      IOPBOOT     CDVDFSV     XFILEIO
	SBIN        ADDDRV      OSDCNF      FILEIO      XSIO2MAN
	LOGO        STDIO       -           CLEARSPU    XMTAPMAN
	IOPBTCONF   SIFMAN      TBIN        UDNL        XMCMAN
	...

Viewing for example the contents of the ROM file ROMVER using /dev/mem:

	# grep -l ROMVER /sys/rom/rom0/file/*/name
	/sys/rom/rom0/file/3/name
	# cd /sys/rom/rom0/file/3
	# dd if=/dev/mem bs=$(cat size) iflag=skip_bytes \
		skip=$(( $(cat data) )) count=1 status=none
	0170EC20030227

For convenience, the ROMVER file is also available directly in sysfs:

	# ls /sys/rom/rom0/version
	date    number  region  type
	# cat /sys/rom/rom0/version/*
	2003-02-27
	0x0170
	Europe
	CEX

The CEX type indicates that it is a retail machine, as opposed to for
example DEX that would be a debug machine.

Signed-off-by: Fredrik Noring <noring@nocrew.org>