MIPS: PS2: IOP: IRQ support
IOP interrupts are not directly available to the kernel. IOP IRQs are
instead serviced by an IRQ relay module, and then forwarded to the
kernel by remote procedure calls (RPCs) via the sub-system interface
(SIF).

The IRQ relay module can also forward IRQs via the SMFLAG (sub-to-main)
register. This is more efficient, but the number of flags is limited.
Currently only RPC forwarding is implemented in the kernel.

Signed-off-by: Fredrik Noring <noring@nocrew.org>