MIPS: PS2: FB: Reset the Graphics Synthesizer drawing environment
Various parameters are used to draw Graphics Synthesizer primitives, for
example texture information and drawing attributes set by the PRIM
register[1]. These parameters are called the drawing environment. The
environment remains in effect for multiple primitives until it is reset.

Some environment registers such as XYOFFSET_1 and XYOFFSET_2 represent
the same function and can be chosen with the CTXT primitive attribute.

This driver is currently only using the first drawing environment.

References:

[1] "GS User's Manual", version 6.0, Sony Computer Entertainment Inc.,
    pp. 47-49.

Signed-off-by: Fredrik Noring <noring@nocrew.org>