MIPS: PS2: FB: fb_settile() with font stored as palette indexed textures
Storing the font map locally in Graphics Synthesizer memory is very
efficient when drawing. A 4-bit palette indexed texture is used so
that both the background and foreground colours can be updated without
reinstalling the font map.

Before using converted data from host-to-local or local-to-local
transmissions as texture or colour lookup tables for the first time,
the texture buffer must be disabled with the TEXFLUSH register.

The register write waits for the completion of the current drawing
process, and then disables the texture data read to the texture page
buffer. Any data can be written to it. A drawing process succeeding
the write to the register is started after the texture page buffer
is disabled[1].

References:

[1] "GS User's Manual", version 6.0, Sony Computer Entertainment Inc.,
    pp. 43, 51, 55, 63, 77, 131.

Signed-off-by: Fredrik Noring <noring@nocrew.org>