MIPS: PS2: IOP: Avoid linking already linked library modules
As a simplification the IOP module linker assumes that modules with
exported libraries are resident. The primary effect is that it refuses
to link modules exporting libraries with the same names more than once,
even if the they may have been unlinked in the IOP. This is because the
IOP module linker maintains its own list of exported libraries, rather
than asking the IOP about them, which remains to be implemented. For now
it is assumed that modules do not unlink themselves.

The IOP module linker also assumes that if a module exports a library it
has the same name as the module. The simplifies resolving dependencies.
In general, and with most ROM modules, the module name is not the same
as the exported library names. A single module may export multiple
libraries, but that is currently not not fully supported with IOP
modules for Linux, due to dependency resolving limitations.

Signed-off-by: Fredrik Noring <noring@nocrew.org>