MIPS: PS2: GS: Primitive and texel coordinate transformations
The frame buffer coordinate system is the pixel drawing space, with
integer coordinates.

The primitive coordinate system is the drawing space used for vertices,
with 4-bit fractional parts.

The texel coordinate system is used for textures, with 4-bit fractional
parts, centered on the position where the fractional parts are 0.5[1].

References:

[1] "GS User's Manual", version 6.0, Sony Computer Entertainment Inc.,
    pp. 23-24,  28.

Signed-off-by: Fredrik Noring <noring@nocrew.org>