MIPS: PS2: IOP: Patch IOP LOADFILE module to improve compatibility
The ROM file rom0:OSDCNF is not available with certain models such as
SCPH-10000, SCPH-15000, DTL-H10000, TOOL, the Namco System 246 archade
machine, and the PS2 emulators in the PS3 and the PS4.

The main task of rom0:OSDCNF for the Linux kernel is to replace LOADFILE
with the XLOADFILE remote procedure call (RPC) service having a working
LoadModuleBuffer function.

This change removes rom0:OSDCNF by patching the existing LOADFILE
service. The RPC switch case is extended from 6 to 7 entries and its
unused entry function is overwritten with a 128 byte patch calling
LoadModuleBuffer and StartModule in rom0:MODLOAD.

rom0:UDNL is also removed as it is no longer necessary. It is used
to pull together all the IOPRP (IOP replacement) images specified as
reset arguments, before selecting the IOP modules from IOPBTCONF with
the highest version number.

The reset argument "rom0:UDNL rom0:OSDCNF" typically loads 31 modules,
whereas the empty argument "" only loads 24 modules that are the default
ones. The 7 modules no longer loaded are prefixed with '-' in the list
of loaded modules:

	 cdvd_driver
	 cdvd_ee_driver
	 dmacman
	 Exception_Manager
	 FILEIO_service
	 Heap_lib
	 Interrupt_Manager
	 IO/File_Manager
	 IOP_SIF_manager
	 IOP_SIF_rpc_interface
	 LoadModuleByEE
	-mcman_cex
	-mcserv
	 Module_Manager
	 Moldule_File_loader
	-multitap_manager
	 Multi_Thread_Manager
	-padman
	 RebootByEE
	-rmman2
	 ROM_file_driver
	-rspu2_driver
	 secrman_for_cex
	-sio2man
	 ssbus_service
	 Stdio
	 SyncEE
	 System_C_lib
	 System_Memory_Manager
	 Timer_Manager
	 Vblank_service